+++ src/utils/index.ts
@@ -1,3 +1,6 @@
 export * from './request';
 export * from './env';
 export * from './color';
+export * from './dateUtil';
+export * from './tsxHelper';
+export * from './props';