+++ CHANGELOG.md
@@ -6,6 +6,21 @@
 * feat: :sparkles: login ([a99611c](https://github.com/Mask-MJ/naive-admin/commit/a99611c))
 * feat: :sparkles: router ([11e38d4](https://github.com/Mask-MJ/naive-admin/commit/11e38d4))
 * feat: init ([a925c40](https://github.com/Mask-MJ/naive-admin/commit/a925c40))
+* feat(header): 新增面包屑导航 ([c0d8567](https://github.com/Mask-MJ/naive-admin/commit/c0d8567))
+* feat(router): 新增路由404 ([7e3dd74](https://github.com/Mask-MJ/naive-admin/commit/7e3dd74))
+* feat(tabs): 新增tab页 ([4312974](https://github.com/Mask-MJ/naive-admin/commit/4312974))
+* refactor(commit): 修改 git commit 校验 ([4c433bd](https://github.com/Mask-MJ/naive-admin/commit/4c433bd))
+
+
+
+## 0.0.0 (2023-12-16)
+
+* feat: :sparkles: add theme ([f00cdf7](https://github.com/Mask-MJ/naive-admin/commit/f00cdf7))
+* feat: :sparkles: init ([1f4e6e4](https://github.com/Mask-MJ/naive-admin/commit/1f4e6e4))
+* feat: :sparkles: layouts ([5c92177](https://github.com/Mask-MJ/naive-admin/commit/5c92177))
+* feat: :sparkles: login ([a99611c](https://github.com/Mask-MJ/naive-admin/commit/a99611c))
+* feat: :sparkles: router ([11e38d4](https://github.com/Mask-MJ/naive-admin/commit/11e38d4))
+* feat: init ([a925c40](https://github.com/Mask-MJ/naive-admin/commit/a925c40))
 * feat(router): 新增路由404 ([7e3dd74](https://github.com/Mask-MJ/naive-admin/commit/7e3dd74))
 * feat(tabs): 新增tab页 ([4312974](https://github.com/Mask-MJ/naive-admin/commit/4312974))
 * refactor(commit): 修改 git commit 校验 ([4c433bd](https://github.com/Mask-MJ/naive-admin/commit/4c433bd))